Refactor filter part in AirportPAge
