+++ frontend/src/pages/AirportPage/AirportPage.tsx
@@ -1,26 +1,21 @@
 import './AirportPage.css'
-import {ChangeEvent, FormEvent, useEffect, useState} from "react";
+import {useEffect, useState} from "react";
 import {Airport} from "../../types/model/dataType.ts";
 import axios from "axios";
 import {BasicUser} from "../../types/auth/userType.ts";
 import {capitalizeFirstLetter} from "../../utils/funtioncs.ts";
-import {AirportFilter} from "../../types/enum.ts";
-import {useFetchOptions} from "../../hooks/useFetchOptions.ts";
+import {AirportFilterType} from "../../types/enum.ts";
+import AirportFilter from "./AirportFilter/AirportFilter.tsx";
 
 type AirportPageProps = {
     loggedInUser: BasicUser | null | undefined,
 }
 
 export default function AirportPage({ loggedInUser }: AirportPageProps ) {
-    const { airports } = useFetchOptions();
     const [airportsData, setAirportsData] = useState<Airport[]>([]);
-    const [selectedFilter, setSelectedFilter] = useState<AirportFilter>({
+    const [selectedFilter, setSelectedFilter] = useState<AirportFilterType>({
         airport: undefined
     });
-    const [filter, setFilter] = useState<AirportFilter>({
-        airport: undefined
-    });
-    const [isDisabled, setIsDisabled] = useState<boolean>(true);
 
     const fetchAllAirports = () => {
         if (loggedInUser?.username) {
@@ -36,58 +31,14 @@ export default function AirportPage({ loggedInUser }: AirportPageProps ) {
         fetchAllAirports();
     }, []);
 
-    const handleChange = (event: ChangeEvent<HTMLSelectElement>) => {
-        setFilter({ airport: event.target.value });
-        setIsDisabled(false);
-    }
-
-    const handleReset = () => {
-        setSelectedFilter({ airport: undefined })
-        setIsDisabled(true);
-    }
-
-    const handleSubmit = (e: FormEvent) => {
-        e.preventDefault();
-        setSelectedFilter({ airport: filter.airport })
-    }
-
     const filteredAirportsData = airportsData
         .filter(airport => selectedFilter.airport ? airport.iataCode === selectedFilter.airport : airport);
 
     return (
         <div>
             <h3>Airport Information</h3>
-            <form onSubmit={handleSubmit} className={"filter-form"}>
-                <div className={"dropdown-filter"}>
-                    <label className={"dropdown-label"}>Airport</label>
-                    <select
-                        name={"origin"}
-                        onChange={handleChange}
-                        className={"origin-dropdown"}
-                    >
-                        <option>All</option>
-                        {airports.map((airport) => (
-                            <option key={airport.code} value={airport.code}>
-                                {airport.code + ' - ' + airport.name}
-                            </option>
-                        ))}
-                    </select>
-                </div>
-                <div className={"filter-buttons"}>
-                    <button
-                        type={"reset"}
-                        className={"reset-btn"}
-                        onClick={handleReset}
-                    >Reset
-                    </button>
-                    <button
-                        type={"submit"}
-                        className={"apply-btn"}
-                        disabled={isDisabled}
-                    >Apply
-                    </button>
-                </div>
-            </form>
+            <AirportFilter setSelectedFilter={setSelectedFilter} />
+
             <section>
                 {filteredAirportsData.map(airport => (
                     <div key={airport.id} className={"airport-card"}>
